requirements.txt
discord.py
aoihttp
websockets
chardet